Backend/requirements.txt
fastapi[standard]
mysql-connector-python
python-jose[cryptography]
bcrypt<4.0.0
passlib[bcrypt]